# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: NovelTea API
  description: |
    Self-hosted, offline-first long-form writing app. Every client keeps a full local replica; this API is a synchronisation point and a compile engine, not where the work lives.

    A few things about the shape of the responses that are easy to get wrong from the schema alone:

    - **A resource the caller may not see returns `404`, never `403`** — a 403 would confirm it exists. Unauthenticated is `401`, so a client knows whether refreshing its token would help.
    - **`POST /projects/{id}/sync` answers `200` even when some changes conflict.** A conflict is an ordinary outcome, reported per change in the response body (`applied` / `conflicts`), not an HTTP error. A `document` conflict never overwrites or merges — the server keeps its version and the client's is preserved as a sibling "conflict copy" (`conflictCopyId`); see `ConflictReason` for why an individual change was rejected.
    - **`GET /projects/{id}/sync` can answer `resyncRequired: true`.** The client's cursor points into history the server can no longer explain (purged, or the project was restored from backup) and must discard it, rebuild from `GET /binder` plus documents, and resume at the returned `latestId`.
...
    - **Compiling is asynchronous.** `POST /projects/{id}/compile` returns a job id immediately; poll `GET /compile-jobs/{id}` until its status is terminal, then `GET /compile-jobs/{id}/download`.
...
  version: v1
servers:
- url: /
...
security:
- bearerAuth: []
...
paths:
  /api/v1/snapshots/{snapshotId}/restore:
    post:
      tags:
      - Snapshots
      summary: Restore a document to a snapshot
      description: "Itself undoable: the pre-restore state is captured automatically\
        \ first, so restoring to the wrong version does not lose the newer one. Refuses\
        \ a stale baseVersion so it cannot clobber an edit made on another device\
        \ meanwhile."
      operationId: restore
      parameters:
      - name: snapshotId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestoreRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: integer
                  format: int64
  /api/v1/projects:
    get:
      tags:
      - Projects
      operationId: list
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      tags:
      - Projects
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Project"
  /api/v1/projects/{projectId}/sync:
    get:
      tags:
      - Sync
      summary: Pull the change feed since a cursor
      description: |-
        Returns rows appended since `since` (exclusive), each carrying the entity's current state. Always answers 200; the response body, not the status code, carries the two conditions a client must check:

        - **`resyncRequired: true`** — the cursor points into history the server can no longer explain, either because it was purged by retention or because the project was restored from an older backup (`syncEpoch` changed). `changes` is empty; the client must discard its cursor, rebuild from `GET /binder` plus documents, and resume pulling from `latestId`.
        - **`hasMore: true`** — more rows exist past this page; pull again from `latestId` before considering the client caught up.

        `latestId` is the highest id *actually served*, never the feed's true maximum — advancing past unserved rows would skip them permanently.
      operationId: pull
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: since
        in: query
        required: false
        schema:
          type: integer
          format: int64
          default: 0
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 200
      - name: epoch
        in: query
        required: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PullResponse"
    post:
      tags:
      - Sync
      summary: Push local changes
      description: |-
        Applies each change independently and always answers 200 — a per-change conflict is a normal, expected outcome of concurrent editing, not a request-level failure. The response separates `applied` from `conflicts` so a client can retry only what actually failed.

        A `document` write that loses a conflict is never merged and never dropped: the server keeps its own version, and the client's is preserved as a sibling "conflict copy" binder item, whose id comes back as `conflictCopyId` on the `ConflictRecord` — non-null there means the author's text survived elsewhere and needs manual reconciliation via the Conflicts endpoints. Tree writes (`binder_item`) are last-write-wins instead; only document content gets a conflict copy.

        See `ConflictReason` for the `reason` values: `version_mismatch` (stale `baseVersion` — the conflict-copy case above), `duplicate_create`, `entity_missing`, `invalid_request`, and `not_implemented` for entity types this push path does not yet handle.
      operationId: push
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PushResponse"
  /api/v1/projects/{projectId}/restore:
    post:
      tags:
      - Projects
      summary: Restore a soft-deleted project
      description: "The one place a deleted project is still visible to its owner\
        \ — everywhere else, deletion makes it answer 404 like it never existed."
      operationId: restore_1
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Project"
...
  /api/v1/documents/{documentId}/snapshots:
    get:
      tags:
      - Snapshots
      operationId: list_1
      parameters:
      - name: documentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SnapshotSummary"
    post:
      tags:
      - Snapshots
      operationId: capture
      parameters:
      - name: documentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaptureRequest"
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
                  format: uuid
...
  /api/v1/binder-items/{itemId}/restore:
    post:
      tags:
      - Binder
      operationId: restore_2
      parameters:
      - name: itemId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
...
  /api/v1/projects/{projectId}:
    get:
      tags:
      - Projects
      summary: Get a project
      description: "A project the caller does not own answers 404, identically to\
        \ one that does not exist — a 403 would confirm it exists to someone who was\
        \ never granted access."
      operationId: get
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "404":
          description: "Not found, or not visible to the caller."
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      tags:
      - Projects
      operationId: delete
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
    patch:
      tags:
      - Projects
      operationId: update
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Project"
...
  /api/v1/snapshots/{snapshotId}:
    get:
      tags:
      - Snapshots
      operationId: get_1
      parameters:
      - name: snapshotId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SnapshotDetail"
    delete:
      tags:
      - Snapshots
      operationId: delete_2
      parameters:
      - name: snapshotId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
...
  /api/v1/projects/{projectId}/purge:
    delete:
      tags:
      - Projects
      summary: Permanently destroy a project and everything in it
      description: "Separate verb and path from delete on purpose: not something a\
        \ client should reach by accident. Irreversible."
      operationId: purge
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
...
components:
  schemas:
    RestoreRequest:
      type: object
      properties:
        baseVersion:
          type: integer
          format: int64
    CreateProjectRequest:
      type: object
      properties:
        title:
          type: string
        settings:
          $ref: "#/components/schemas/JsonNode"
    JsonNode:
      type: object
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ownerId:
          type: string
          format: uuid
        title:
          type: string
        settings:
          $ref: "#/components/schemas/JsonNode"
        documentCount:
          type: integer
          format: int32
        wordCount:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
    ChangeRequest:
      type: object
      properties:
        entityType:
          type: string
        entityId:
          type: string
          format: uuid
        op:
          type: string
        baseVersion:
          type: integer
          format: int64
        data:
          $ref: "#/components/schemas/JsonNode"
    PushRequest:
      type: object
      properties:
        since:
          type: integer
          format: int64
        changes:
          type: array
          items:
            $ref: "#/components/schemas/ChangeRequest"
    AppliedChange:
      type: object
      properties:
        entityId:
          type: string
          format: uuid
        entityType:
          type: string
        version:
          type: integer
          format: int64
    ConflictRecord:
      type: object
      properties:
        entityId:
          type: string
          format: uuid
        entityType:
          type: string
        reason:
          type: string
          description: "One of ConflictReason: version_mismatch (stale baseVersion\
            \ — see conflictCopyId), duplicate_create, entity_missing, invalid_request,\
            \ or not_implemented for an entity type this push path does not yet handle."
          example: version_mismatch
        conflictCopyId:
          type: string
          format: uuid
          description: "The binder_item holding the client's rejected document version,\
            \ when one was created. Non-null means the author's text was preserved\
            \ elsewhere and needs manual reconciliation via the Conflicts endpoints;\
            \ null means nothing was at risk (e.g. a tree write, which is last-write-wins)."
        serverVersion:
          type: integer
          format: int64
        detail:
          type: string
    PushResponse:
      type: object
      properties:
        applied:
          type: array
          items:
            $ref: "#/components/schemas/AppliedChange"
        conflicts:
          type: array
          items:
            $ref: "#/components/schemas/ConflictRecord"
        latestId:
          type: integer
          format: int64
...
    CaptureRequest:
      type: object
      properties:
        label:
          type: string
        automatic:
          type: boolean
...
    UpdateProjectRequest:
      type: object
      properties:
        title:
          type: string
        settings:
          $ref: "#/components/schemas/JsonNode"
...
    SnapshotDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        documentId:
          type: string
          format: uuid
        label:
          type: string
        automatic:
          type: boolean
        wordCount:
          type: integer
          format: int32
        content:
          $ref: "#/components/schemas/JsonNode"
        createdAt:
          type: string
          format: date-time
    ChangeRecord:
      type: object
      properties:
        id:
          type: integer
          format: int64
        entityType:
          type: string
        entityId:
          type: string
          format: uuid
        op:
          type: string
        deviceId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        data:
          $ref: "#/components/schemas/JsonNode"
    PullResponse:
      type: object
      properties:
        changes:
          type: array
          items:
            $ref: "#/components/schemas/ChangeRecord"
        latestId:
          type: integer
          format: int64
          description: Highest change id actually served in this response — never
            the feed's true maximum. A client that advanced past unserved rows would
            skip them permanently.
        hasMore:
          type: boolean
          description: More rows exist past this page; pull again from latestId.
        resyncRequired:
          type: boolean
          description: "The client's cursor points into history the server can no\
            \ longer explain (purged by retention, or the project was restored from\
            \ an older backup). `changes` is empty; the client must rebuild from GET\
            \ /binder plus documents and resume at latestId."
        syncEpoch:
          type: integer
          format: int64
...
    SnapshotSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        documentId:
          type: string
          format: uuid
        label:
          type: string
        automatic:
          type: boolean
        wordCount:
          type: integer
          format: int32
        createdByDeviceId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      description: "Access token from /auth/register, /auth/login, /auth/refresh or\
        \ /auth/pair. 15-minute lifetime by default (noveltea.auth.access-token-ttl)."
      scheme: bearer
      bearerFormat: JWT